+++ src/position/template/app-token.template.position-fetcher.ts
@@ -236,7 +236,7 @@ export abstract class AppTokenTemplatePositionFetcher<T extends Contract, V exte
           const displayProps = {
             label: await this.getLabel(displayPropsStageParams),
             labelDetailed: await this.getLabelDetailed(displayPropsStageParams),
-            secondarylabel: await this.getSecondaryLabel(displayPropsStageParams),
+            secondaryLabel: await this.getSecondaryLabel(displayPropsStageParams),
             tertiaryLabel: await this.getTertiaryLabel(displayPropsStageParams),
             images: await this.getImages(displayPropsStageParams),
             statsItems: await this.getStatsItems(displayPropsStageParams),